requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
idna==3.3
    # via requests
python-dateutil==2.8.2
    # via
    #   sosi-api
    #   sosi-etherscan (setup.py)
python-decouple==3.5
    # via
    #   sosi-api
    #   sosi-etherscan (setup.py)
requests==2.26.0
    # via
    #   sosi-api
    #   sosi-etherscan (setup.py)
six==1.16.0
    # via python-dateutil
sosi-api @ git+ssh://git@github.com/ronrest/sosi_api.git@v0.0.1
    # via sosi-etherscan (setup.py)
urllib3==1.26.7
    # via requests
